Another night, another case. The only thing different today is the city-
wide mask festival: An offering to the sea after the autumn storms have passed. It will make catching the culprit harder, but not impossible. You are good at finding the pieces that other people hide, turning them into a picture. And so, you enter your crime scene, this time a brothel ...